+++ helpers/helpers_xxx_global.js
@@ -1,7 +1,7 @@
 ﻿'use strict';
-//19/06/25
+//26/06/25
 
-/* exported loadUserDefFile, addGlobValues, globFonts, globSettings*/
+/* exported loadUserDefFile, addGlobValues, globFonts, globSettings, globNoSplitArtist */
 
 include('helpers_xxx.js');
 /* global folders:readable */
@@ -18,7 +18,7 @@ function loadUserDefFile(def) {
 		if (data) {
 			const handleList = new FbMetadbHandleList();
 			const skipCheckKeys = ['_description', '_usage'];
-			if (def._type === 'TF' || def._type === 'Query' || def._type === 'Font' || def._type === 'Setting') {
+			if (def._type === 'TF' || def._type === 'Query' || def._type === 'Font' || def._type === 'Setting' || def._type === 'Set' || def._type === 'Array') {
 				for (const key in data) {
 					if (Object.hasOwn(def, key)) {
 						def[key] = data[key];
@@ -70,6 +70,15 @@ function loadUserDefFile(def) {
 									);
 									def[key].size = 10;
 								}
+							} else if (def._type === 'Set' || def._type === 'Array') {
+								if (!def[key]) {
+									fb.ShowPopupMessage(
+										'There has been an error trying to parse the setting:\n' + key + ' (' + def._type + ' type)' +
+										'\n\nValue is empty. It must be filled with at least and empty array.'
+										, 'Loading config from ' + def._file
+									);
+								}
+								if (def._type === 'Set') { def[key] = new Set(def[key]); }
 							}
 						}
 					} else { bSave = true; }
@@ -109,7 +118,14 @@ function loadUserDefFile(def) {
 	} else { addGlobValues(def._type); bSave = true; }
 	if (bSave) {
 		const { _type, _file, ...rest } = def; // eslint-disable-line no-unused-vars
-		_save(_file, JSON.stringify(rest, (key, value) => { return (value instanceof RegExp ? value.toSource() : value); }, '\t').replace(/\n/g, '\r\n'));
+		_save(
+			_file,
+			JSON.stringify(rest, (key, value) => {
+				if (value instanceof RegExp) { return value.toSource(); }
+				else if (value instanceof Set) { return [...value]; }
+				else { return value; }
+			}, '\t').replace(/\n/g, '\r\n')
+		);
 	}
 }
 
@@ -125,7 +141,7 @@ function loadUserDefFile(def) {
  * @returns {void}
  */
 function addGlobValues(type) {
-	const _t = (tag) => !tag.includes('%') && !tag.includes('$') ? '%' + tag+ '%' : tag;
+	const _t = (tag) => !tag.includes('%') && !tag.includes('$') ? '%' + tag + '%' : tag;
 	switch (type) {
 		case 'TF':
 			globTags.title = '$ascii($lower($trim($replace(' + _t(globTags.titleRaw) + ',\'\',,`,,’,,´,,-,,\\,,/,,:,,$char(34),))))'; // Takes ~1 sec on 80K tracks;
@@ -302,4 +318,13 @@ const globSettings = {
 	bTooltip: true,
 	bLogToFile: false,
 	instanceManager: 'v1'
+};
+
+// No-Split Artist: user replaceable with a presets file at folders.data
+const globNoSplitArtist = {
+	_type: 'Set',
+	_file: folders.userPresetsGlobal + 'globNoSplitArtist.json',
+	_description: 'List of artists which should not be split (usually due to comma usage).',
+	_usage: 'Artists can be added manually or at relevant scripts via menus.',
+	list: ['10,000 Maniacs', 'Crosby, Stills, Nash & Young', 'Crosby, Stills & Nash', 'Our Friend, Surrender']
 };